Split observations module into submodules
