+++ gym_sts/spaces/observations/__init__.py
@@ -1,0 +1,1 @@
+from .observations import OBSERVATION_SPACE, Observation  # noqa: F401
