+++ gym_sts/spaces/observations/components/base.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class ObsComponent(ABC):
+    @abstractmethod
+    def serialize(self):
+        raise RuntimeError("Not implemented")
