+++ gym_sts/spaces/observations/components/persistent.py
@@ -1,0 +1,67 @@
+from pydantic import parse_obj_as
+
+from gym_sts.spaces import constants
+from gym_sts.spaces.observations import serializers, types, utils
+
+from .base import ObsComponent
+from .map import MapStateObs
+
+
+class PersistentStateObs(ObsComponent):
+    def __init__(self, state: dict):
+        # Sane defaults
+        self.hp = 0
+        self.max_hp = 0
+        self.gold = 0
+        self.potions = []
+        self.relics = []
+        self.deck = []
+        self.keys = {}
+        self.map = MapStateObs()
+
+        if "game_state" in state:
+            game_state = state["game_state"]
+            self.hp = game_state["current_hp"]
+            self.max_hp = game_state["max_hp"]
+            self.gold = game_state["gold"]
+            self.potions = parse_obj_as(list[types.Potion], game_state["potions"])
+            self.relics = parse_obj_as(list[types.Relic], game_state["relics"])
+            self.deck = parse_obj_as(list[types.Card], game_state["deck"])
+            self.map = MapStateObs(state)
+
+            if "keys" in game_state:
+                self.keys = game_state["keys"]
+
+    def serialize(self) -> dict:
+        health = serializers.serialize_health(self.hp, self.max_hp)
+        gold = utils.to_binary_array(self.gold, constants.LOG_MAX_GOLD)
+
+        potions = [0] * constants.NUM_POTION_SLOTS
+
+        for i, potion in enumerate(self.potions):
+            potions[i] = constants.ALL_POTIONS.index(potion.id)
+
+        _relics = [False] * constants.NUM_RELICS
+        for relic in self.relics:
+            _relics[constants.ALL_RELICS.index(relic.id)] = True
+        relics = [int(relic) for relic in _relics]
+
+        deck = serializers.serialize_cards(self.deck)
+
+        _keys = [False] * constants.NUM_KEYS
+        for i, key in enumerate(["ruby", "emerald", "sapphire"]):
+            if key in self.keys:
+                _keys[i] = self.keys[key]
+        keys = [int(key) for key in _keys]
+
+        response = {
+            "health": health,
+            "gold": gold,
+            "potions": potions,
+            "relics": relics,
+            "deck": deck,
+            "keys": keys,
+            "map": self.map.serialize(),
+        }
+
+        return response
