+++ gym_sts/spaces/observations/serializers.py
@@ -1,25 +1,7 @@
 from gym_sts.spaces import constants
+from gym_sts.spaces.observations import types
 
-from . import types
-
-
-def to_binary_array(n: int, digits: int) -> list[int]:
-    array = [0] * digits
-
-    idx = 0
-    n_copy = n
-    while n_copy > 0:
-        if idx >= digits:
-            raise ValueError(
-                f"{n} is too large to represent with {digits} binary digits"
-            )
-
-        n_copy, r = divmod(n_copy, 2)
-        if r > 0:
-            array[idx] = 1
-        idx += 1
-
-    return array
+from .utils import to_binary_array
 
 
 def serialize_cards(cards: list[types.Card]) -> list[int]:
